Issue#5images&aesthetics (#9)
* `<div>` in Index.cshtml has been partially filled with sample text

MilkMan img added, for `README.md`-- logo image

* MilkMan Martin logo added, placeholder for carousel on home page created. New README.md needed on other issues. Layout reformating almost complete.

Refactored some of the HTML in the Index.cshtml -- still working on it.

* BerryPie limitededition milk image and chocolate milk images addes. Carousel has been implmented with placeholder text.

Carousel needs to be made smaller and thinner without damaging the integrity of the images.

* Working on resizing the products on the main page into a carousel for aesthetics. No resizing done has been permanent, still working on proper sizing.

`site.css` only thing that was changed was the @media was at min-width - 768px now it is max-width - 768px

* More aesthetic issues cleared up for the time being, buttons have been pushed closer together (ISSUE- the buttons are not visible and the carousel does not have automatic playback please fix)

Figured out how to use the `site.css`, restructured the website so that it's not so hard on the eyes.

Logo was removed from the title and added to the navbar instead- Later on down the road a MOO `button` will be added for Martin.

* Title, Name, Company Name changed from Milkmans Milk to Milkman Farms...less times the word"milk" is included.

Small aesthetic modifications solved. No MP3-Moo updated.
Trees.png was addes for a background image behind the carousel but has not been applied to the CSS.

* Create CodingGuidelines.md (#8)

closes #6 
Code and Structure basics created and made.

* ProductController made, New Issue to be created for testing.
-Use Bootstrap to make individual cards for the different flavors of milk provided.
-Add to Cart Button needed
- Cart in NavBar needed.

* Product View Created not implemented

Layout adds a new portion to the nav bar

* Product name placed