New cart controller (#14)
* Basic cards made using Bootstrap in the Product index. Images were also added!

* Names added to the cards in the Product index View.

Placeholders replaced with temporary description, trying to get css for the cards working to better space out the cards on the product view.

Product cards are not spaced correctly. Please space out slightly.

Added placeholder pricing until I can get sample data

* in the process of making a database to replace, placeholders in the index of the Products page.

Add to cart functionality to be made--

* CartController made, IActionResult method name changed to Cart... there are already two 'indexes" in the other controllers.

* Cart view name changed back to index lol turns out it's less confusing that way !!

* Listing out what I am strategizing at the moement using the <p>, new cartcontroller sucesssfully made and added to the nav bar.

preferred location for viewing the cart would be the far right side of the nav bar.

View made for the CartController, add and remove functions needed for the cart in addition to a small calulator function used with adding or removing an item from the cart.

* Mascot name changed in Readme, it was wrong.

Card made in the Index for the Cart, Product Class made, with properties.

A Data folder was also created in hopes to connect a database to this and implement CRUD functionality!! eCommerceDatabaseContext class made but not implemented. Pull Request~~

* comment added for next issue