Updated the Cart/Index, so that the shipping and totals updaate as the View updates.
UpdateQuantity remade so that we can update order. OrderDetails summary has not been created that will be for advanced updates over break.

Validation for CartItems in the Cart View- if there are no items in the cart then an caution message will appear. There is currently no limit on Quantity

Card on the Cart has been updated, pull request to be made soon