Tidying&validation (#26)
* Migrations made, updating registration pages so that the new users can be sent and updated as well.

updating aesthetics and turning the registration page form in to a card not done.

* Tidying up the program and updating aesthetics. comments added, TempData added to product LoginMessage.

Using a boolean we could see if the login information provided matches the email and password in the database. If it does then a sucess alert will appear/ else/ we are just going to assume that it is invaild for now, until we add a forgot password option-----
 some of the sample data is missing passwords so forgotten password to update should be added to issues

* adding more comments to each file.

* one full form for the cart will appear once a product is added

ViewProducts button made on the Cart/Index page so that if there are no cartitems then the button will appear.

full circle now. Pull Request